SCHOOL-18 unit test
